requirements.txt
# Transfer Learning MVP - Python dependencies
# Install: pip install -r requirements.txt

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Data and evaluation
Pillow>=9.0.0
requests>=2.28.0
scikit-learn>=1.2.0

# Inference API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
